Add JDK class custom serializer test
